requirements.txt
# Streamlit Cloud 배포용 의존성
streamlit>=1.28.0
gspread>=5.11.0
google-auth>=2.23.0
pandas>=2.0.0
plotly>=5.17.0
python-dateutil>=2.8.2
requests>=2.31.0
